refactor(web): change folder structure
