+++ apps/web/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,3 @@
-import { handlers } from "../../../../shared/auth/auth";
+import { handlers } from "../../../../src/shared/auth/auth";
 
 export const { GET, POST } = handlers;
