+++ apps/web/app/login/page.tsx
@@ -1,34 +1,11 @@
-"use client";
+import { LoginPage } from "@/views/login";
+import type { Metadata } from "next";
 
-import { useSession } from "next-auth/react";
-import GoogleButton from "./_components/GoogleButton";
+export const metadata: Metadata = {
+  title: "Time Align | 로그인",
+  description: "일정관리의 시작, Time Align",
+};
 
-export default function TestAuthPage() {
-  const { data: session } = useSession();
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        flexDirection: "column",
-        justifyContent: "center",
-        alignItems: "center",
-        height: "100vh",
-      }}
-    >
-      <GoogleButton />
-      <div style={{ width: 300, marginTop: 20 }}>
-        <pre
-          style={{
-            whiteSpace: "pre-wrap",
-            wordWrap: "break-word",
-            overflowWrap: "break-word",
-            maxWidth: "100%",
-          }}
-        >
-          {JSON.stringify(session, null, 2)}
-        </pre>
-      </div>
-    </div>
-  );
+export default function Page() {
+  return <LoginPage />;
 }
