+++ apps/web/app/page.tsx
@@ -1,10 +1,5 @@
+import { HomePage } from "@/views/home";
 import type { Metadata } from "next";
-import { FAB } from "../shared/components/FAB";
-import { ScheduleCard } from "../shared/components/ScheduleCard";
-import { css } from "../styled-system/css";
-import { Calendar } from "./_components/Calendar";
-import { Tabs } from "./_components/Tabs";
-import { TopBanner } from "./_components/TopBanner";
 
 export const metadata: Metadata = {
   title: "Time Align | 홈",
@@ -12,33 +7,5 @@ export const metadata: Metadata = {
 };
 
 export default function Page() {
-  return (
-    <section
-      className={css({
-        display: "flex",
-        flexDirection: "column",
-      })}
-    >
-      <TopBanner />
-      <Calendar />
-      <Tabs />
-      <ul
-        className={css({
-          display: "flex",
-          flexDirection: "column",
-          py: "18px",
-          px: "22px",
-          gap: "12px",
-          bg: "#EFF2F5",
-        })}
-      >
-        {Array.from({ length: 5 }).map((_, index) => (
-          <li key={`schedule-card-${index.toString()}`}>
-            <ScheduleCard />
-          </li>
-        ))}
-      </ul>
-      <FAB />
-    </section>
-  );
+  return <HomePage />;
 }
