+++ apps/web/app/profile/page.tsx
@@ -1,3 +1,4 @@
+import { ProfilePage } from "@/views/profile";
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -6,9 +7,5 @@ export const metadata: Metadata = {
 };
 
 export default function Page() {
-  return (
-    <section>
-      <div>Profile</div>
-    </section>
-  );
+  return <ProfilePage />;
 }
