+++ apps/web/src/shared/api/endpoints/app/app.ts
