+++ apps/web/src/shared/api/endpoints/auth/auth.ts
