+++ apps/web/src/shared/auth/auth.ts
