+++ apps/web/src/shared/components/BottomNavigation.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { css } from "../../styled-system/css";
+import { css } from "../../../styled-system/css";
 
 const NAV_ITEMS = [
   {
