+++ apps/web/src/shared/components/ScheduleCard.tsx
@@ -1,4 +1,4 @@
-import { css } from "../../styled-system/css";
+import { css } from "../../../styled-system/css";
 
 export function ScheduleCard() {
   return (
