+++ apps/web/src/shared/exception/APIError.ts
