+++ apps/web/src/shared/exception/BaseError.ts
