+++ apps/web/src/views/documents/index.tsx
@@ -1,0 +1,7 @@
+export function DocumentsPage() {
+  return (
+    <section>
+      <div>Documents</div>
+    </section>
+  );
+}
