+++ apps/web/src/views/profile/index.tsx
@@ -1,0 +1,7 @@
+export function ProfilePage() {
+  return (
+    <section>
+      <div>Profile</div>
+    </section>
+  );
+}
