+++ apps/web/src/widgets/home/Calendar.tsx
@@ -1,4 +1,4 @@
-import { css } from "../../styled-system/css";
+import { css } from "styled-system/css";
 
 export function Calendar() {
   return (
