+++ apps/web/src/widgets/login/GoogleButton.tsx
