+++ apps/web/src/widgets/schedule/ScheduleProgressDivider.tsx
