+++ apps/web/src/widgets/schedule/ScheduleProgressItem.tsx
