+++ apps/web/src/widgets/schedule/Tabs.tsx
